refactored the font code
